Rename routes, read chat context from file, clear local storage on Result page
